requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
openai==1.82.0
requests==2.32.3
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gunicorn==21.2.0